# swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "User Management Service",
    "description": "User Management Service Swagger Documentation"
  },
  "host": "localhost:1234",
  "basePath": "/users",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/": {
      "x-swagger-router-controller": "user",
      "get": {
        "operationId": "findAll",
        "description": "find all users",
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/SearchResponse"
            }
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/{id}": {
      "x-swagger-router-controller": "user",
      "get": {
        "operationId": "findById",
        "description": "Find User By Id",
        "parameters": [
          {
            "name": "id",
            "type": "string",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/SearchResponse"
            }
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "put": {
        "operationId": "update",
        "description": "Find & Update User By Id",
        "parameters": [
          {
            "name": "id",
            "type": "string",
            "in": "path",
            "required": true
          },
          {
            "name": "body",
            "description": "request body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GeneralResponse"
            }
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      },
      "delete": {
        "operationId": "delete",
        "description": "Delete User By Id",
        "parameters": [
          {
            "name": "id",
            "type": "string",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GeneralResponse"
            }
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/create": {
      "x-swagger-router-controller": "user",
      "post": {
        "operationId": "create",
        "description": "create user",
        "parameters": [
          {
            "name": "title",
            "description": "Creates a sale record",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UserDetails"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/GeneralResponse"
            }
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/elastic": {
      "x-swagger-router-controller": "user",
      "get": {
        "operationId": "es_searchAll",
        "description": "find all users in elastic",
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/SearchResponse"
            }
          },
          "default": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        }
      }
    },
    "/swagger": {
      "x-swagger-pipe": "swagger_raw"
    }
  },
  "definitions": {
    "SuccessResponse": {
      "required": [
        "message"
      ],
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "SearchResponse": {
      "required": [
        "user"
      ],
      "properties": {
        "user": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "age": {
                "type": "number"
              }
            }
          }
        }
      }
    },
    "UserDetails": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "description": "user name"
        },
        "age": {
          "type": "number",
          "description": "age of user"
        }
      },
      "required": [
        "name",
        "user"
      ]
    },
    "UserRequest": {
      "required": [
        "name",
        "age"
      ],
      "properties": {
        "name": {
          "type": "string"
        },
        "age": {
          "type": "number"
        }
      }
    },
    "GeneralResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "number",
          "description": "returns 1 if successful"
        },
        "description": {
          "type": "string",
          "description": "a short comment"
        }
      },
      "required": [
        "success",
        "description"
      ]
    },
    "ErrorResponse": {
      "required": [
        "message"
      ],
      "properties": {
        "message": {
          "type": "string"
        }
      }
    }
  }
}